IOS-5027 Moved viewmodel into separate file
